Move get job log file method to base config.
Remove duplicate base service method.
