+++ test/unit/services/base/config_test.py
@@ -1,3 +1,5 @@
+from unittest.mock import patch
+
 from mash.mash_exceptions import MashConfigException
 from mash.services.base_config import BaseConfig
 from pytest import raises
@@ -108,3 +110,13 @@ class TestBaseConfig(object):
     def test_get_notification_subject(self):
         subject = self.empty_config.get_notification_subject()
         assert subject == '[MASH] Job Status Update'
+
+    def test_get_log_dir(self):
+        assert self.config.get_log_directory() == '/var/log/mash/'
+        assert self.empty_config.get_log_directory() == '/var/log/mash/'
+
+    @patch.object(BaseConfig, 'get_log_directory')
+    def test_get_job_log_file(self, mock_get_log_dir):
+        mock_get_log_dir.return_value = '/var/log/mash/'
+        assert self.empty_config.get_job_log_file('1234') == \
+            '/var/log/mash/jobs/1234.log'
